@@ -14,9 +14,9 @@
     "@mantine/notifications": "^5.0.2",
     "@portabletext/react": "^1.0.6",
     "@sanity/image-url": "^1.0.1",
-    "next": "12.2.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^13.5.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-toggle-dark-mode": "^1.1.0"
   },
   "devDependencies": {
